+++ Sources/Controllers/ALKReplyController.swift
@@ -72,8 +72,6 @@ class ALKReplyController: UIViewController, Localizable {
 
     let configuration: ALKConfiguration
     let messageKey: String
-    let userId: String?
-    let groupId: NSNumber?
 
     private let attachmentView = ALKAttatchmentView(frame: .zero)
 
@@ -143,9 +141,7 @@ class ALKReplyController: UIViewController, Localizable {
         return view
     }()
 
-    init(userId: String?, groupId: NSNumber?, messageKey: String, configuration: ALKConfiguration) {
-        self.userId = userId
-        self.groupId = groupId
+    init(messageKey: String, configuration: ALKConfiguration) {
         self.messageKey = messageKey
         self.configuration = configuration
         super.init(nibName: nil, bundle: nil)