+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The changelog for [ApplozicSwift](https://github.com/AppLozic/ApplozicSwift). Al
 
 ## [Unreleased]
 
+### Enhancements
+- [CM-402] Date picker support in form template.
+
 ## [5.10.2] - 2020-10-06
 
 ### Fixes